ml/serving/requirements.txt
# Inference server dependencies
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
google-cloud-storage>=2.10.0
